More templates for building scanners: late_riprel_to_vtbl_m(vtbl_name) skip_call_within_distance_and_capture_target<calltarget_out, max_distance>
scan and acquire gamelocal getplayerbyidx and peermask_for_player

network message stuff

"Consequence free" mh_reload_decl. This is a crazy feature that works by setting up an alternate stack/cpu context that yields back to the main thread when a Error/FatalError is raised or when an SEH exception is caught. This may not save the user from a crash in all cases, and it is possible that some memory corruption could occur as a result of this, but there is a greater chance that you will be able to continue after a failed mh_reload_decl than previously where the chance was 0%. Despite this, i'd still recommend restarting the game after failing a decl reload. the results following the failure may be unpredictable

HOW IT WORKS:

An early feature of mh was intercepting id's FatalError and Error functions and displaying a messagebox with the error message. Originally, the game would just crash with no info. While it's great to have this info, when reloading a decl it sucks to have the game terminate. There seemed to be no other way though, since the FatalError and Error functions are noreturn if i were to just return to the caller the engine would likely crash anyway or behave weirdly.

So instead, we save the current execution context for our thread at the start of mh_reload_decl and set up an alternate stack to be used for local variables. We scope this execution context globally, and in the error handlers we switch contexts back, passing along the error message. In case of a segfault we set up an seh handler

Add scanner for handlereliable message